Gaussian Cross Entropy Implemented
This commit baselines uncertainty measure for classification task.
Particularly for segmentation. It introduces gaussian noise over
the final logits predictions using a regressed squared-sigma channel.
Multiple cross-entropy losses are calculated using monte-carlo sims,
the result is then averaged across the number of sims and is treated
as the loss.

This commit takes