Implement TwinDeepLab
This experiment intends to understand the interplay between depth and
rgb for small obstacle detection. It implements a dual CNN framwork,
where one is responsible for extracting image features, and the other
is responsible to extract depth gradients. This is done by regressing
an image confidence and depth confidence channel separately. These
confidences are then concatenated, converted to probs via softmax, and
then used to weight the feature map tensors of both the networks.

bug-fixes
 - saving weight files
 - visualises softmax values
 - visualises image confidence
 - visualises depth confidence