Assimilate Lidar and RGB for SO Segmentation
Basic implementation of sparse Lidar RGB camera fusion for small
obstacle segmentation. Lidar is represented as a single depth channel.
This is concatenated as the 4th channel in a 4 channel sparse representation
of RGBD. Thinking of improving it by adding another channel that encodes
presence or absence of Lidar value for a pixel.